+++ packages/core/index.ts
@@ -4,7 +4,7 @@ export {
   EventCollectionEmitter,
   ICollectionItem,
 } from './engine/event-collection';
-export { IKites, KitesInstance, IKitesOptions } from './engine/kites';
+export { IKites, KitesInstance, IKitesOptions } from './engine/kites-instance';
 export { KitesExtension, ExtensionDefinition, ExtensionOptions } from './extensions/extensions';
 
-export * from './engine/main';
+export * from './engine/kites-factory';
